requirements.txt
psycopg2-binary
pandas
openpyxl
faker
reportlab
python-dotenv